split huge parsing test into a new file
